migrate create dao related types out of typings file
